Create project-independent class
